stack and queue data
This is program in which you can perform push and pop operation. It take user input for perform operation which you want. there is two part of program, first part show how to data add and delete in list and second part show that how to data add and delete in queue When you will be out of moves the program will be exit